requirements.txt
Django==1.9.6
django-braces==1.10.0
django-cors-headers==1.3.0
django-oauth-toolkit==0.10.0
djangorestframework==3.5.3
oauthlib==1.0.3
PyMySQL==0.7.9
six==1.10.0
